LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/font-display/font-display-feature-policy-reporting.tentative-expected.txt

CONSOLE MESSAGE: line 68: ReferenceError: Can't find variable: ReportingObserver
Tests if the correct number of violation reports are generated and each report corresponds to this feature. 4 reports should be created out of the 6 options below (reports for all except for 'fallback' and 'optional').

not-set	auto	block	swap	fallback	optional
a	a	a	a	a	a

Harness Error (FAIL), message = ReferenceError: Can't find variable: ReportingObserver

NOTRUN font-display-late-swap Report Format 
